v38r6

+++ Event/EventPacker/src/component/ReadPackedDst.cpp
@@ -187,7 +187,7 @@ StatusCode ReadPackedDst::execute() {
       put( caloHypos, name + m_postfix );
       processLinks( caloHypos, version );
       getFromBlob<LHCb::PackedCaloHypo> ( caloHypos->hypos() , blobs );
-      getFromBlob<int>                  ( caloHypos->refs()  , blobs );
+      getFromBlob<long long>            ( caloHypos->refs()  , blobs );
 
     } else if ( LHCb::CLID_PackedProtoParticles  == classID ) {
 
@@ -195,7 +195,7 @@ StatusCode ReadPackedDst::execute() {
       put( protos, name + m_postfix );
       processLinks( protos, version );
       getFromBlob<LHCb::PackedProtoParticle> ( protos->protos() , blobs );
-      getFromBlob<int>                       ( protos->refs()   , blobs );
+      getFromBlob<long long>                 ( protos->refs()   , blobs );
       getFromBlob<std::pair<int,int> >       ( protos->extras() , blobs );
 
     } else if ( LHCb::CLID_PackedRecVertices     == classID ) {
@@ -204,8 +204,8 @@ StatusCode ReadPackedDst::execute() {
       put( recVertices, name + m_postfix );
       processLinks( recVertices, version );
       getFromBlob<LHCb::PackedRecVertex> ( recVertices->vertices() , blobs );
-      getFromBlob<int>                   ( recVertices->refs()    , blobs );
-      getFromBlob<std::pair<int,int> >   ( recVertices->extras() , blobs );
+      getFromBlob<long long>             ( recVertices->refs()     , blobs );
+      getFromBlob<std::pair<int,int> >   ( recVertices->extras()   , blobs );
 
     } else if ( LHCb::CLID_PackedTwoProngVertices     == classID ) {
 
@@ -213,8 +213,8 @@ StatusCode ReadPackedDst::execute() {
       put( recVertices, name + m_postfix );
       processLinks( recVertices, version );
       getFromBlob<LHCb::PackedTwoProngVertex> ( recVertices->vertices() , blobs );
-      getFromBlob<int>                        ( recVertices->refs()    , blobs );
-      getFromBlob<std::pair<int,int> >        ( recVertices->extras() , blobs );
+      getFromBlob<long long>                  ( recVertices->refs()     , blobs );
+      getFromBlob<std::pair<int,int> >        ( recVertices->extras()   , blobs );
 
     } else if ( LHCb::CLID_RecSummary == classID ) {
